@@ -6,11 +6,11 @@
 
     <groupId>org.example</groupId>
     <artifactId>estudoAlgoritmos</artifactId>
-    <version>1.7.0</version>
+    <version>1.8.0</version>
 
     <properties>
-        <maven.compiler.source>19</maven.compiler.source>
-        <maven.compiler.target>19</maven.compiler.target>
+        <maven.compiler.source>22</maven.compiler.source>
+        <maven.compiler.target>22</maven.compiler.target>
         <retrofit.version>2.6.0</retrofit.version>
     </properties>
 
@@ -27,7 +27,7 @@
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
-                <version>0.8.8</version>
+                <version>0.8.12</version>
                 <executions>
                     <execution>
                         <id>default-prepare-agent</id>
